Commit after git fiasco. Changes to be committed:     modified:   conf.php     modified:   core/auth.php     new file:   core/langman.php     new file:   core/langman/policy-description-schema-v01.rng     new file:   core/langman/text-schema-v1.rng     new file:   core/langman/title-schema-v1.rng     modified:   core/policy.php     modified:   core/unifunctions.php     modified:   core/userman.php
Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Thu Oct 2 03:24:56 2014 +0500

    New method getTitleNamesXML was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Thu Oct 2 03:14:14 2014 +0500

    New method sumTitleNames was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Thu Oct 2 01:47:37 2014 +0500

    New method getTextNamesXML was added. Some other changes.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Sep 30 02:45:39 2014 +0500

    New method sumTextNames was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Sep 27 17:23:29 2014 +0500

    Some changes in getTextSectionsXML and getTitleSectionsXML.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Sep 27 16:06:21 2014 +0500

    New method getTitleSectionsXML was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Fri Sep 26 19:35:48 2014 +0500

    New method sumTitleSections was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Fri Sep 26 18:21:31 2014 +0500

    New method getTextSectionsXML was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Sep 23 03:49:59 2014 +0500

    New method sumTextSections was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Mon Sep 22 19:55:24 2014 +0500

    New method getPolicyDescById was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Wed Sep 17 20:05:05 2014 +0500

    New method groupPolicyList was added to LangMan. Method policyList was deleted from Policy.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Sep 16 18:24:27 2014 +0500

    New method getTitleById was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Sep 16 12:50:27 2014 +0500

    New method getAllTitlesBySectionXML was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Sep 16 12:33:16 2014 +0500

    New method getTitlesBySectionXML was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Sep 16 12:10:22 2014 +0500

    New method sumTitlesBySection was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Mon Sep 15 18:24:28 2014 +0500

    New method getTextsBySection was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Mon Sep 15 11:07:56 2014 +0500

    New method getTextsBySectionXML was added. Some changes were made inside method sumTextsBySection.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Sep 9 13:56:50 2014 +0500

    New method sumTextsBySection was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Sep 6 18:25:52 2014 +0500

    New method getTextById was added. Output of the method getText was changed.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Fri Sep 5 04:09:36 2014 +0500

    New method getAllTextsBySectionXML was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Aug 30 13:11:30 2014 +0500

    New method getTitleSection was added. New parameter $code was added to methods addText and addTitle.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Aug 30 04:02:51 2014 +0500

    New method getText was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Thu Aug 28 14:34:59 2014 +0500

    New methods setLang and getTitle were added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Aug 26 16:12:02 2014 +0500

    New method updateText was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Aug 26 13:22:32 2014 +0500

    New method updateTitle was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Aug 26 04:55:58 2014 +0500

    New method delText was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Aug 26 04:52:00 2014 +0500

    New method addText was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Aug 26 04:21:18 2014 +0500

    New method delTitle was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Aug 26 03:12:10 2014 +0500

    New method addTitle was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Mon Aug 25 05:25:22 2014 +0500

    Incoming parameters of delTextName were changed.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Mon Aug 25 05:20:35 2014 +0500

    Incoming parameters of delTitleName were changed.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Mon Aug 25 04:05:35 2014 +0500

    Changed incoming parameters of delTextSection.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Mon Aug 25 03:34:40 2014 +0500

    Changed incoming parameters of delTitleSection.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sun Aug 24 01:54:53 2014 +0500

    New method delTextName.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Aug 23 23:55:53 2014 +0500

    New method addTextName.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Aug 23 23:50:47 2014 +0500

    New method delTitleName.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Aug 23 04:44:58 2014 +0500

    New method addTitleName.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Aug 23 02:58:55 2014 +0500

    New method delTextSection.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Aug 23 02:53:19 2014 +0500

    New method addTextSection.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Fri Aug 22 22:44:29 2014 +0500

    New method delTitleSection was added.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Fri Aug 22 03:30:17 2014 +0500

    Method addTitleSection was added to langman.php. Method pregName40 was added to unifunctions.php.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Thu Aug 21 22:46:41 2014 +0500

    New method delPlugin was added to langman.php. New method pregPlugin was added to unifunctions.php.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Thu Aug 21 20:03:14 2014 +0500

    New method installTexts was added. New schema Relax NG.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Tue Aug 19 18:26:23 2014 +0500

    New method installTitles was added. New Relax NG schema for langman.php.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Thu Aug 7 02:00:46 2014 +0500

    Table _core_langman_about_policy was renamed to _core_langman_policy_description. New method installPolicyDesc was added to core module langman.php. Some changes for updating of policy descriptions were made in core module policy.php.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sun Jul 27 01:36:28 2014 +0500

    In auth.php session variable core_auth_langid was replaced with core_auth_lang. In unifunctions.php function authLangId was replaced with authLang.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sun Jul 27 01:09:01 2014 +0500

    Method langList was added. It allows to get a list of all available languages as DOMDocument object.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Jul 26 22:37:22 2014 +0500

    Method delLang was added. It deletes specified language with all related data.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Jul 26 14:17:12 2014 +0500

    New method authLangId was added to unifunctions.php. Method addLang was added to new core module langman.php.

Author: Alexei Muzarov <azexmail@gmail.com>
Date:   Sat Jul 26 00:45:33 2014 +0500

    New system constant MECCANO_DEF_LANG was added to conf.php. Method createUser in userman.php was changed to enable support of multi-language. Also methods userLogin and getSession in auth.php were changed to enable support of multi-language.

+++ core/policy.php
@@ -22,13 +22,6 @@ class Policy {
     private static function setErrExp($exp) {
         self::$errexp = $exp;
     }
-    
-    private static function pregName($name) {
-        if (is_string($name) && preg_match('/^[a-zA-Z\d_]{3,30}$/', $name)) {
-            return TRUE;
-        }
-        return FALSE;
-    }
 
     public static function errId() {
         return self::$errid;
@@ -61,6 +54,12 @@ class Policy {
         $newFuncs = array_diff($functions, $dbFuncs);
         // deleting of outdated policies
         foreach ($oldFuncs as $funcId) {
+            self::$dblink->query("DELETE FROM `".MECCANO_TPREF."_core_langman_policy_description` "
+                    . "WHERE `policyid`=$funcId ;");
+            if (self::$dblink->errno) {
+                self::setErrId(ERROR_NOT_EXECUTED);                self::setErrExp('installPolicy: can\'t delete policy description | '.self::$dblink->error);
+                return FALSE;
+            }
             self::$dblink->query("DELETE FROM `".MECCANO_TPREF."_core_policy_access` "
                     . "WHERE `funcid`=$funcId ;");
             if (self::$dblink->errno) {
@@ -122,12 +121,16 @@ class Policy {
     
     public static function delPolicy($name) {
         self::$errid = 0;        self::$errexp = '';
-        if (!self::pregName($name)) {
+        if (!pregPlugin($name)) {
             self::setErrId(ERROR_INCORRECT_DATA);            self::setErrExp('delPolicy: name must be string');
             return FALSE;
         }
         $plugName = self::$dblink->real_escape_string($name);
         $queries = array(
+            "DELETE `d` FROM `".MECCANO_TPREF."_core_langman_policy_description` `d` "
+            . "JOIN `".MECCANO_TPREF."_core_policy_summary_list` `s` "
+            . "ON `s`.`id`=`d`.`policyid` "
+            . "WHERE `s`.`name`='$plugName' ;",
             "DELETE `a` FROM `".MECCANO_TPREF."_core_policy_access` `a` "
             . "JOIN `".MECCANO_TPREF."_core_policy_summary_list` `s` "
             . "ON `s`.`id`=`a`.`funcid` "
@@ -202,7 +205,7 @@ class Policy {
     
     public static function funcAccess($name, $func, $groupid, $access = TRUE) {
         self::$errid = 0;        self::$errexp = '';
-        if (!(is_integer($groupid) || is_bool($groupid)) || !self::pregName($name) || !self::pregName($func)) {
+        if (!(is_integer($groupid) || is_bool($groupid)) || !pregPlugin($name) || !pregPlugin($func)) {
             self::setErrId(ERROR_NOT_EXECUTED);            self::setErrExp('funcAccess: incorect type of incoming parameters');
             return FALSE;
         }
@@ -253,42 +256,9 @@ class Policy {
         return TRUE;
     }
     
-    public static function policyList($name, $groupid) {
-        self::$errid = 0;        self::$errexp = '';
-        if (!self::pregName($name) || !(is_integer($groupid) || is_bool($groupid))) {
-            self::setErrId(ERROR_INCORRECT_DATA);            self::setErrExp('policyList: incorect type of incoming parameters');
-            return FALSE;
-        }
-        $plugName =  self::$dblink->real_escape_string($name);
-        if (is_bool($groupid)) {
-            $qList = self::$dblink->query("SELECT `s`.`func`, `n`.`access` "
-                    . "FROM `".MECCANO_TPREF."_core_policy_summary_list` `s` "
-                    . "JOIN `".MECCANO_TPREF."_core_policy_nosession` `n` "
-                    . "ON `s`.`id`=`n`.`funcid` "
-                    . "WHERE `s`.`name`='$plugName' ;");
-        }
-        else {
-            $qList = self::$dblink->query("SELECT `s`.`func`, `a`.`access` "
-                    . "FROM `".MECCANO_TPREF."_core_policy_summary_list` `s` "
-                    . "JOIN `".MECCANO_TPREF."_core_policy_access` `a` "
-                    . "ON `s`.`id`=`a`.`funcid` "
-                    . "WHERE `s`.`name`='$plugName' "
-                    . "AND `a`.`groupid`=$groupid ;");
-        }
-        if (self::$dblink->errno) {
-            self::setErrId(ERROR_NOT_EXECUTED);            self::setErrExp('policyList: something went wrong | '.self::$dblink->error);
-            return FALSE;
-        }
-        if (!self::$dblink->affected_rows) {
-            self::setErrId(ERROR_NOT_FOUND);            self::setErrExp('policyList: name or group don\'t exist');
-            return FALSE;
-        }
-        return $qList;
-    }
-    
     public static function checkAccess($name, $func) {
         self::$errid = 0;        self::$errexp = '';
-        if (!self::pregName($name) || !self::pregName($func)) {
+        if (!pregPlugin($name) || !pregPlugin($func)) {
             self::setErrId(ERROR_INCORRECT_DATA);            self::setErrExp('checkAccess: check incoming parameters');
             return FALSE;
         }
